pyproject.toml
[project]
name = "review-rating-prediction"
version = "0.1.0"
description = "Children's books rating prediction using ML"
requires-python = "==3.11.*"
dependencies = [
    "matplotlib==3.8.4",
    "seaborn==0.12.2",
    "ruff==0.1.14",
    "sentence-transformers==2.2.2",
    "torch==2.3.0",
    "joblib==1.4.2",
    "mlflow==2.13.0",
    "pandas==2.2.2",
    "numpy==1.26.4",
    "scikit-learn==1.5.0",
    "xgboost==2.0.3",
    "fastapi==0.111.0",
    "uvicorn==0.28.1",
    "pytest==8.2.2",
    "pyyaml==6.0.1",
    "setuptools==69.5.1",
    "prefect==2.18.1",
    "griffe==0.38.1",
    "pydantic==1.10.14"
]

[build-system]
requires = ["setuptools>=69.0.0", "wheel"]
build-backend = "setuptools.build_meta"